cmd: move debug commands to a separate debug.rs file
This follows #1061 for consistency. The resulting file is of a similar size to
operation.rs.
